Update PegasusPlugin to remove stale schema files when preparing schemas for publication.
There was a bug where stale schema files were being packaged into the
data-template jars. There is a flag that allows us to delete the stale
files but this caused a different problem where the tasks would never be
up to date, which would cause other issues if we were running a server
in a hot reload mode.

The fix is to use a sync task instead of a delete and copy tasks.

* Fix PegasusPlugin to use the sync task when moving the files to the
build/mainSchemas for data template preparation
* Create test to validate stale files are deleted
* Create test to validate mainCopySchema task will be up to date when no
changes happen